% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algo.R
\name{genetic_algorithm}
\alias{genetic_algorithm}
\title{Genetic Algorithm for Student-Project Assignment}
\usage{
genetic_algorithm(
  cost_matrix,
  pop_size = 100,
  generations = 500,
  mutation_rate = 0.1
)
}
\arguments{
\item{cost_matrix}{A square numeric matrix where entry \code{[i, j]}
represents the cost of assigning student \code{i} to project \code{j}.}

\item{pop_size}{An integer specifying the number of individuals (assignments)
in each generation. Default is 100.}

\item{generations}{An integer giving the number of generations to run the
genetic algorithm. Default is 500.}

\item{mutation_rate}{A numeric value (between 0 and 1) giving the probability
of mutation. Default is 0.1.}
}
\value{
A list with two elements:
\describe{
\item{cost}{The total cost of the best assignment found.}
\item{assignment}{An integer vector giving the assigned project for
each student. Each position i in the vector corresponds to student i,
and the value at that position indicates the assigned project. }
}
}
\description{
Solves the student-project allocation problem using a genetic algorithm.
The goal is to find an assignment of students to projects that minimizes
the total cost.
}
\examples{
generate_matrix <- function(n) {
matrix(apply(matrix(1:n, n, n, byrow = TRUE), 1, sample), 
nrow = n, byrow = TRUE)}
cost <- generate_matrix(4)
result <- genetic_algorithm(cost)
print(result$cost)
print(result$assignment)

}
